+++ examples/box/index.json
@@ -1,45 +1,115 @@
 {
   "app": {
-    "version": "sunmao/v1",
-    "kind": "Application",
-    "metadata": {
-      "name": "some App"
-    },
+    "version": "example/v1",
+    "metadata": { "name": "box", "description": "box" },
     "spec": {
       "components": [
         {
-          "id": "text",
-          "type": "core/v1/text",
-          "properties": {
-            "value": {
-              "raw": "text",
-              "format": "plain"
-            }
-          },
+          "id": "test_btn",
+          "type": "chakra_ui/v1/button",
+          "properties": { "text": { "raw": "Click", "format": "plain" } },
           "traits": [
             {
-              "type": "test/v1/timeout",
+              "type": "core/v1/state",
+              "properties": { "key": "count", "initialValue": 0 }
+            },
+            {
+              "type": "core/v1/event",
               "properties": {
-                "value": "{{input.value + Math.random()}}"
+                "handlers": [
+                  {
+                    "type": "onClick",
+                    "componentId": "test_btn",
+                    "method": {
+                      "name": "setValue",
+                      "parameters": {
+                        "key": "count",
+                        "value": "{{ test_btn.count > 0 ? 0 : test_btn.count + 1 }}"
+                      }
+                    },
+                    "wait": {},
+                    "disabled": false
+                  }
+                ]
               }
             }
           ]
         },
         {
-          "id": "input",
-          "type": "test/v1/input",
+          "id": "box_container",
+          "type": "chakra_ui/v1/box",
           "properties": {
-            "defaultValue": "foo"
+            "my": 8,
+            "p": 8,
+            "display": "flex",
+            "bg": "{{ test_btn.count % 2 ? 'teal' : 'white' }}",
+            "border": "1px solid",
+            "borderColor": "{{ test_btn.count % 2 ? 'white' : 'teal' }}",
+            "borderRadius": 8,
+            "h": 200
           },
           "traits": []
         },
         {
-          "id": "tester",
-          "type": "test/v1/tester",
+          "id": "box_child_1",
+          "type": "chakra_ui/v1/box",
           "properties": {
-            "text": "{{text.result}}"
+            "mr": 4,
+            "flex": 1,
+            "bg": "{{ test_btn.count % 2 ? 'white' : 'teal' }}",
+            "color": "{{ test_btn.count % 2 ? 'teal' : 'white' }}",
+            "display": "flex",
+            "alignItems": "center",
+            "justifyContent": "center",
+            "borderRadius": 4
           },
-          "traits": []
+          "traits": [
+            {
+              "type": "core/v1/slot",
+              "properties": { "container": { "id": "box_container", "slot": "content" } }
+            }
+          ]
+        },
+        {
+          "id": "text1",
+          "type": "core/v1/text",
+          "properties": { "value": { "raw": "in box child 1", "format": "plain" } },
+          "traits": [
+            {
+              "type": "core/v1/slot",
+              "properties": { "container": { "id": "box_child_1", "slot": "content" } }
+            }
+          ]
+        },
+        {
+          "id": "box_child_2",
+          "type": "chakra_ui/v1/box",
+          "properties": {
+            "flex": 1,
+            "bg": "{{ test_btn.count % 2 ? 'white' : 'teal' }}",
+            "color": "{{ test_btn.count % 2 ? 'teal' : 'white' }}",
+            "display": "flex",
+            "alignItems": "center",
+            "justifyContent": "center",
+            "borderRadius": 4
+          },
+          "traits": [
+            {
+              "type": "core/v1/slot",
+              "properties": { "container": { "id": "box_container", "slot": "content" } }
+            }
+          ]
+        },
+        {
+          "id": "text2",
+          "type": "core/v1/text",
+          "properties": { "value": { "raw": "in box child 2", "format": "plain" } },
+          "traits": [
+            {
+              "type": "core/v1/slot",
+              "properties": { "container": { "id": "box_child_2", "slot": "content" } }
+            }
+          ]
         }
       ]
     }
