+++ packages/editor/src/main.tsx
@@ -1,3 +1,4 @@
+import { StrictMode } from 'react';
 import ReactDOM from 'react-dom';
 import { RegistryInterface } from '@sunmao-ui/runtime';
 import { sunmaoChakraUILib } from '@sunmao-ui/chakra-ui-lib';
@@ -42,5 +43,10 @@ export default function renderApp(options: Options = {}) {
   traits.forEach(t => registry.registerTrait(t));
   modules.forEach(m => registry.registerModule(m));
 
-  ReactDOM.render(<Editor />, container);
+  ReactDOM.render(
+    <StrictMode>
+      <Editor />
+    </StrictMode>,
+    container
+  );
 }
