+++ packages/runtime/index.html
@@ -31,14 +31,19 @@
         ReactDOM.unmountComponentAtNode(rootEl);
         const { App, registry } = initSunmaoUI({ libs: [sunmaoChakraUILib] });
         const { app, modules = [] } = example.value;
+        window.registry = registry;
         modules.forEach(m => {
           registry.registerModule(m);
         });
         ReactDOM.render(
           React.createElement(
-            ChakraProvider,
+            React.StrictMode,
             null,
-            React.createElement(App, { options: app }, null)
+            React.createElement(
+              ChakraProvider,
+              null,
+              React.createElement(App, { options: app }, null)
+            )
           ),
           rootEl
         );
