+++ packages/runtime/src/components/_internal/ImplWrapper/ImplWrapperMain.tsx
@@ -191,8 +191,6 @@ export const ImplWrapperMain = React.forwardRef<HTMLDivElement, ImplWrapperProps
   }
 );
 
-// ImplWrapperMain.whyDidYouRender = true;
-
 // This hook will only run unmount function when unmount, not every time when unmount function changes.
 const useDidUnmount = (fn: Function) => {
   const fnRef = useRef(fn);
